Modification des tables

+++ src/src/models.py
@@ -2,8 +2,8 @@ from django.db import models
 
 class Film(models.Model):
     titre = models.CharField(max_length=100)
-    note_spectateur = models.IntegerField()
-    note_presse = models.IntegerField()
+    note_spectateur = models.FloatField()
+    note_presse = models.FloatField()
     nombre_commentaires = models.IntegerField()
 
     def __str__(self):
@@ -15,7 +15,7 @@ class Film(models.Model):
 class Commentaire(models.Model):
     pseudo = models.TextField(null=True)
     commentaire = models.TextField()
-    note = models.IntegerField()
+    note = models.FloatField()
     titre_film = models.CharField(max_length=100, null=True)
 
 
@@ -28,7 +28,7 @@ class Commentaire(models.Model):
 class NLP(models.Model):
     model = models.CharField(max_length=255)
     film = models.ForeignKey(Film, on_delete=models.CASCADE)
-    note = models.IntegerField()
+    note = models.FloatField()
 
     def __str__(self):
         return f"{self.model} - {self.film.titre}"
